duplicated views removed and docs endpoint path fixed
