+++ EnviroMonitorWeb/EnviroMonitorWeb/urls.py
@@ -15,10 +15,9 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
-from docs.views import schema_view
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^', include('api.urls')),
-    url(r'^docs/',  schema_view),
+
 ]
