+++ EnviroMonitorWeb/api/models.py
@@ -1,13 +1,7 @@
 from django.core.urlresolvers import reverse
 from django_extensions.db.fields import AutoSlugField
-from django.db.models import *
 from django.conf import settings
-from django.contrib.contenttypes.fields import GenericForeignKey
-from django.contrib.contenttypes.models import ContentType
-from django.contrib.auth import get_user_model
-from django.contrib.auth import models as auth_models
 from django.db import models as models
-from django_extensions.db import fields as extension_fields
 
 
 class Station(models.Model):
@@ -82,7 +76,7 @@ class Project(models.Model):
 
     # Fields
     name = models.CharField(max_length=255)
-    slug = extension_fields.AutoSlugField(populate_from='name', blank=True)
+    slug = AutoSlugField(populate_from='name', blank=True)
     created = models.DateTimeField(auto_now_add=True, editable=False)
     last_updated = models.DateTimeField(auto_now=True, editable=False)
     project_website = models.URLField()
