--- EnviroMonitorWeb/api/templates/api/metering_detail.html
@@ -1,23 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_metering_list' %}">Metering Listing</a></p>
-<table class="table">
-<tr><td>created</td><td>{{ object.created }}</td></tr>
-<tr><td>pm01</td><td>{{ object.pm01 }}</td></tr>
-<tr><td>pm25</td><td>{{ object.pm25 }}</td></tr>
-<tr><td>pm10</td><td>{{ object.pm10 }}</td></tr>
-<tr><td>temp_out1</td><td>{{ object.temp_out1 }}</td></tr>
-<tr><td>temp_out2</td><td>{{ object.temp_out2 }}</td></tr>
-<tr><td>temp_out3</td><td>{{ object.temp_out3 }}</td></tr>
-<tr><td>hum_out1</td><td>{{ object.hum_out1 }}</td></tr>
-<tr><td>hum_out2</td><td>{{ object.hum_out2 }}</td></tr>
-<tr><td>hum_out3</td><td>{{ object.hum_out3 }}</td></tr>
-<tr><td>temp_int1</td><td>{{ object.temp_int1 }}</td></tr>
-<tr><td>hum_int1</td><td>{{ object.hum_int1 }}</td></tr>
-<tr><td>rssi</td><td>{{ object.rssi }}</td></tr>
-<tr><td>bpress_out1</td><td>{{ object.bpress_out1 }}</td></tr>
-</table>
-<a class="btn btn-primary" href="{{object.get_update_url}}">Edit Metering</a>
-
-{% endblock %}
