--- EnviroMonitorWeb/api/templates/api/metering_form.html
@@ -1,11 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load crispy_forms_tags %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_metering_list' %}">Metering Listing</a></p>
-<form method="post">
-{% csrf_token %}
-{{form|crispy}}
-<button class="btn btn-primary" type="submit">Submit</button>
-</form>
-{% endblock %}
