--- EnviroMonitorWeb/api/templates/api/metering_list.html
@@ -1,44 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_metering_list' %}">Metering Listing</a></p>
-<table class="table">
-<tr>
-<td>ID</td><td>Link</td>
-    <td>created</td>
-    <td>pm01</td>
-    <td>pm25</td>
-    <td>pm10</td>
-    <td>temp_out1</td>
-    <td>temp_out2</td>
-    <td>temp_out3</td>
-    <td>hum_out1</td>
-    <td>hum_out2</td>
-    <td>hum_out3</td>
-    <td>temp_int1</td>
-    <td>hum_int1</td>
-    <td>rssi</td>
-    <td>bpress_out1</td>
-</tr>
-{% for object in object_list %}
-<tr>
-    <td>{{object.id}}</td>
-    <td><a href="{{object.get_absolute_url}}">{{object}}</a></td>
-    <td>{{ object.created }}</td>
-    <td>{{ object.pm01 }}</td>
-    <td>{{ object.pm25 }}</td>
-    <td>{{ object.pm10 }}</td>
-    <td>{{ object.temp_out1 }}</td>
-    <td>{{ object.temp_out2 }}</td>
-    <td>{{ object.temp_out3 }}</td>
-    <td>{{ object.hum_out1 }}</td>
-    <td>{{ object.hum_out2 }}</td>
-    <td>{{ object.hum_out3 }}</td>
-    <td>{{ object.temp_int1 }}</td>
-    <td>{{ object.hum_int1 }}</td>
-    <td>{{ object.rssi }}</td>
-    <td>{{ object.bpress_out1 }}</td>
-</tr>
-{% endfor %}
-</table><a class="btn btn-primary" href="{% url 'api_metering_create' %}">Create new Metering</a>
-{% endblock %}
