--- EnviroMonitorWeb/api/templates/api/project_detail.html
@@ -1,16 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_project_list' %}">Project Listing</a></p>
-<table class="table">
-<tr><td>name</td><td>{{ object.name }}</td></tr>
-<tr><td>slug</td><td>{{ object.slug }}</td></tr>
-<tr><td>created</td><td>{{ object.created }}</td></tr>
-<tr><td>last_updated</td><td>{{ object.last_updated }}</td></tr>
-<tr><td>project_website</td><td>{{ object.project_website }}</td></tr>
-<tr><td>description</td><td>{{ object.description }}</td></tr>
-<tr><td>logo</td><td>{{ object.logo }}</td></tr>
-</table>
-<a class="btn btn-primary" href="{{object.get_update_url}}">Edit Project</a>
-
-{% endblock %}
