--- EnviroMonitorWeb/api/templates/api/project_form.html
@@ -1,11 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load crispy_forms_tags %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_project_list' %}">Project Listing</a></p>
-<form method="post">
-{% csrf_token %}
-{{form|crispy}}
-<button class="btn btn-primary" type="submit">Submit</button>
-</form>
-{% endblock %}
