--- EnviroMonitorWeb/api/templates/api/project_list.html
@@ -1,30 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_project_list' %}">Project Listing</a></p>
-<table class="table">
-<tr>
-<td>ID</td><td>Link</td>
-    <td>name</td>
-    <td>slug</td>
-    <td>created</td>
-    <td>last_updated</td>
-    <td>project_website</td>
-    <td>description</td>
-    <td>logo</td>
-</tr>
-{% for object in object_list %}
-<tr>
-    <td>{{object.id}}</td>
-    <td><a href="{{object.get_absolute_url}}">{{object}}</a></td>
-    <td>{{ object.name }}</td>
-    <td>{{ object.slug }}</td>
-    <td>{{ object.created }}</td>
-    <td>{{ object.last_updated }}</td>
-    <td>{{ object.project_website }}</td>
-    <td>{{ object.description }}</td>
-    <td>{{ object.logo }}</td>
-</tr>
-{% endfor %}
-</table><a class="btn btn-primary" href="{% url 'api_project_create' %}">Create new Project</a>
-{% endblock %}
