--- EnviroMonitorWeb/api/templates/api/station_detail.html
@@ -1,22 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% block content %}
-<p><a class="btn btn-default" href="{% url 'api_station_list' %}">Station Listing</a></p>
-<table class="table">
-<tr><td>name</td><td>{{ object.name }}</td></tr>
-<tr><td>created</td><td>{{ object.created }}</td></tr>
-<tr><td>last_updated</td><td>{{ object.last_updated }}</td></tr>
-<tr><td>type</td><td>{{ object.type }}</td></tr>
-<tr><td>notes</td><td>{{ object.notes }}</td></tr>
-<tr><td>test</td><td>{{ object.test }}</td></tr>
-<tr><td>position</td><td>{{ object.position }}</td></tr>
-<tr><td>country</td><td>{{ object.country }}</td></tr>
-<tr><td>state</td><td>{{ object.state }}</td></tr>
-<tr><td>county</td><td>{{ object.county }}</td></tr>
-<tr><td>community</td><td>{{ object.community }}</td></tr>
-<tr><td>city</td><td>{{ object.city }}</td></tr>
-<tr><td>district</td><td>{{ object.district }}</td></tr>
-</table>
-<a class="btn btn-primary" href="{{object.get_update_url}}">Edit Station</a>
-
-{% endblock %}
