--- EnviroMonitorWeb/api/templates/base.html
@@ -1,1 +1,0 @@
-<html><head></head><body>{% block content %}Replace this.{% endblock %}</body>
